requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
Flask-Caching
Flask-JWT-Extended
python-dotenv
qrcode
Pillow
email-validator
WTForms
SQLAlchemy
Werkzeug
gunicorn
uwsgi
redis
psycopg2-binary
PyJWT
celery
flower
prometheus-client
alembic
Jinja2
MarkupSafe
itsdangerous
amqp
kombu
billiard
vine
click
blinker